chore: clean TestContext before class
Up to now we used TestContext for all classes in the run.
This cause an issue when some class start emulator/simulator and add them to the context.
Next class will kill them in before class but they will still live in TestContext and thus we will try to capture screenshots after class which will obviously fail.