chore: remove flaky logic in Adb.get_screen()
We called `adb shell screencap -p | perl -pe 's/\\x0D\\x0A/\\x0A/g' >
file.png` because it was faster.
We found on some machines it fails and it works if we use a bit modified
version.
Our assumption was that it happens because of update to macOS Mohave,
but it turned out that that is not the reason.

Fall back to old logic that get screenshot on sdcard and pull it.
It is slower, but looks to be more stable.