fix: debug tests
**Issue Reason**
In latest chrome/driver `actions.double_click(span).perform()` do not cause real double click.

**Impact**
Result is we fail to edit text in elements tab -> failing tests.

**The Fix**
We use `pyautogui` and do real clicks.
In order to do real clicks we need absolute coordiantes of elements so `Chrome.get_absolute_center()` is implemented.

**Notes**
We use strict version of `pyautogui==0.9.39` because as of `pyautogui==0.9.40` it is not compatible with Python 2.